+++ IHP/Controller/NotFound.hs
@@ -16,6 +16,7 @@ import qualified Data.ByteString.Lazy as LBS
 import IHP.HSX.QQ (hsx)
 import qualified System.Directory as Directory
 import IHP.Controller.Context
+import IHP.Controller.Response (respondAndExit)
 
 
 -- | Stops the action execution with a not found message (404) when the access condition is True.
